Implement support for renaming
Internally to the runtime we want to map types more accurately which
requires us to have some kind of indication that a typedef is a pure
rename which has special behaviors (doesn't emit the typedef, merging
using the renamed name behaves as expected instead of overriding).

This is going to be used only within the runtime internals.